fix: improve data type parsing and thread handling in log processing pipeline
- Updated INTEGER and LONG data type parsing to use Double.parseDouble with casting to avoid potential parsing errors
- Refactored output adapter sending logic to use explicit loop instead of deprecated helper method
- Replaced queue polling with blocking take() operations in parser and transformer threads
- Added proper thread interruption handling and cleanup in processing loops
- Improved exception handling with correct interruption propagation and logging

+++ src/main/java/org/keinus/logparser/application/config/ConfigManagementService.java
@@ -374,8 +374,8 @@ public class ConfigManagementService {
 
         try {
             return switch (dataType.toUpperCase()) {
-                case "INTEGER" -> Integer.parseInt(value.trim());
-                case "LONG" -> Long.parseLong(value.trim());
+                case "INTEGER" -> (int) Double.parseDouble(value.trim());
+                case "LONG" -> (long) Double.parseDouble(value.trim());
                 case "BOOLEAN" -> Boolean.parseBoolean(value.trim());
                 case "DOUBLE" -> Double.parseDouble(value.trim());
                 default -> value;
